# api/openapi/v1/keyip.yaml
openapi: "3.0.3"
info:
  title: KeyIP-Intelligence API
  version: "1.0.0"
  description: |
    KeyIP-Intelligence is an AI-driven intellectual property lifecycle management platform
    designed for the OLED organic materials industry. It integrates molecular-level chemical
    understanding with patent law intelligence to support patent mining, infringement
    monitoring, portfolio optimization, and lifecycle management.

    ## Authentication

    All API endpoints require Bearer token authentication unless otherwise specified.
    Tokens are issued via the configured identity provider (Keycloak / OIDC-compatible).

    ## Pagination

    List endpoints support page-based pagination via `page` and `page_size` query parameters.
    Responses include pagination metadata.

    ## Error Handling

    All errors follow a unified structure:
    ```json
    {
      "code": "NOT_FOUND",
      "message": "The specified resource does not exist"
    }
    ```
  contact:
    name: KeyIP-Intelligence Open Source Community
    url: https://github.com/turtacn/KeyIP-Intelligence
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.keyip.example.com/api/v1
    description: Production server
  - url: https://staging-api.keyip.example.com/api/v1
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/patents/{patentId}/lifecycle:
    get:
      tags: [Lifecycle]
      summary: Get patent lifecycle
      operationId: getPatentLifecycle
      parameters:
        - $ref: "#/components/parameters/PatentPathId"
      responses:
        "200":
          description: Patent lifecycle
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/patents/{patentId}/lifecycle/advance:
    post:
      tags: [Lifecycle]
      summary: Advance lifecycle phase
      operationId: advanceLifecyclePhase
      parameters:
        - $ref: "#/components/parameters/PatentPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdvancePhaseRequest"
      responses:
        "200":
          description: Phase advanced
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/patents/{patentId}/milestones:
    get:
      tags: [Lifecycle]
      summary: List milestones
      operationId: listMilestones
      parameters:
        - $ref: "#/components/parameters/PatentPathId"
      responses:
        "200":
          description: Milestone list
          content:
            application/json:
              schema:
                type: object

    post:
      tags: [Lifecycle]
      summary: Add milestone
      operationId: addMilestone
      parameters:
        - $ref: "#/components/parameters/PatentPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMilestoneRequest"
      responses:
        "201":
          description: Milestone added
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/patents/{patentId}/fees:
    get:
      tags: [Lifecycle]
      summary: List fees
      operationId: listFees
      parameters:
        - $ref: "#/components/parameters/PatentPathId"
      responses:
        "200":
          description: Fee list
          content:
            application/json:
              schema:
                type: object

    post:
      tags: [Lifecycle]
      summary: Record fee payment
      operationId: recordFee
      parameters:
        - $ref: "#/components/parameters/PatentPathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordFeeRequest"
      responses:
        "201":
          description: Fee recorded
          content:
            application/json:
              schema:
                type: object
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/patents/{patentId}/timeline:
    get:
      tags: [Lifecycle]
      summary: Get patent timeline
      operationId: getPatentTimeline
      parameters:
        - $ref: "#/components/parameters/PatentPathId"
      responses:
        "200":
          description: Patent timeline
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/deadlines/upcoming:
    get:
      tags: [Lifecycle]
      summary: Get upcoming deadlines
      description: Returns all deadlines due within the specified lookahead window.
      operationId: getUpcomingDeadlines
      parameters:
        - name: days
          in: query
          description: Lookahead window in days (default 30, max 365).
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 365
      responses:
        "200":
          description: Upcoming deadline list
          content:
            application/json:
              schema:
                type: object
  /api/v1/lifecycle/{id}/annuities/calculate:
    post:
      tags: [Lifecycle]
      summary: Calculate annuities
      operationId: calculateAnnuities
      parameters:
        - $ref: "#/components/parameters/LifecycleId"
      responses:
        "200":
          description: Annuity calculation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  patent_id:
                    type: string
                  annuities:
                    type: array
                    items:
                      type: object
                  total:
                    type: number
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/lifecycle/{id}/annuities/budget:
    get:
      tags: [Lifecycle]
      summary: Get annuity budget
      operationId: getAnnuityBudget
      parameters:
        - $ref: "#/components/parameters/LifecycleId"
      responses:
        "200":
          description: Annuity budget
          content:
            application/json:
              schema:
                type: object
                properties:
                  patent_id:
                    type: string
                  budget_items:
                    type: array
                    items:
                      type: object
                  total_amount:
                    type: number
                  fee_count:
                    type: integer
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/lifecycle/{id}/legal-status/sync:
    post:
      tags: [Lifecycle]
      summary: Sync legal status
      operationId: syncLegalStatus
      parameters:
        - $ref: "#/components/parameters/LifecycleId"
      responses:
        "200":
          description: Legal status synced
          content:
            application/json:
              schema:
                type: object
                properties:
                  patent_id:
                    type: string
                  phase:
                    type: string
                  status:
                    type: string
                  synced_at:
                    type: string
                    format: date-time
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/lifecycle/{id}/calendar/export:
    get:
      tags: [Lifecycle]
      summary: Export calendar
      operationId: exportCalendar
      parameters:
        - $ref: "#/components/parameters/LifecycleId"
      responses:
        "200":
          description: Calendar events
          content:
            application/json:
              schema:
                type: object
                properties:
                  patent_id:
                    type: string
                  events:
                    type: array
                    items:
                      type: object
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------------------------------------------------------------------------
  # Collaboration - Workspaces
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PatentPathId:
      name: patentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    LifecycleId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.

    # -------------------------------------------------------------------------
    # Health
    # -------------------------------------------------------------------------
...
    AdvancePhaseRequest:
      type: object
      required: [target_phase]
      properties:
        target_phase:
          type: string
        notes:
          type: string
    AddMilestoneRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        type:
          type: string
    RecordFeeRequest:
      type: object
      required: [fee_type, amount]
      properties:
        fee_type:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        paid_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        description:
          type: string

    # -------------------------------------------------------------------------
    # Collaboration
    # -------------------------------------------------------------------------
...